Update 2021-10-14.md
made some changes--let me know if you have any questions! I'm also wondering if we need to mention the letter is used by the VA Mobile app. Will it always be used by that app, and only that app? If not, we could just say this: 

We've added the Benefits Summary Dependent Letter to the VA Letter Generator API. This letter includes the following information about a dependent's benefits: 
 - Claim number
 - Military service objects, such as the Veteran's disability status and up to 3 of their most recent periods of service