Create 2023-04-04
This release note was published under the Patient Health API 4/04/2023. It was realized that this change applies to both the Patient Health and the Clinical Health API, so this pull request is for getting the same release note that was published 4/4/23 under Patient Health to appear under the Clinical Health API release notes section of the devportal as well.