Create 2023-05-01
This release note was published under the Patient Health API 5/01/2023. It was realized that this change applies to both the Patient Health and the Clinical Health API, so this pull request is for getting the same release note that was published 5/1/23 under Patient Health to appear under the Clinical Health API release notes section of the devportal as well.