API-44751 Update SendToLPB Workflow
The changes in this commit aim to improve the workflow for sending content
to the Lighthouse Platform Backend (LPB).  The primary goal is to handle
deleted files before changed files. This alteration addresses a potential
issue where non-technical users might delete and then recreate release notes
with the same name or date through the Github UI, resulting in data loss.

In order to implement this change the steps for getting and sending deleted
files have been added prior to the steps for getting and sending changed files.
Additionally the changed_files.txt file is now being processed after the
deleted_files.txt.  This will prevent accidental deletion of newly uploaded
release notes with the same name as a previously deleted release note.