API-44751 Update Checkout and Diff Logic
The checkout logic has been modified to always use the main branch
and the diff calculation now writes the result directly to a file.
This allows for easier usage in subsequent steps of the workflow
and improves performance by avoiding the need to output the diff
as a step variable.

Changes have also been made to the changed_files.sh script, which
now assumes that the base branch is main when running locally, and
simplifies the logic for classifying the changed and deleted files.