API-44751 Testing environment contents (#232)
* API-44751 Testing environment contents

* fixed things that were not shell commands into variables?

* changed name of an existing release note

* restored real file content and name; explicitely git rm on test file

* deleted an existing release note after making a backup

* restored existing real file; added a new test file to test git diff script locally

* testing a git status way of identifying changed files

* the git status approach may not work

* added an old file and renamed and modified an old file

* Add script to analyze unique changes in the current Git branch

This commit introduces a new shell script that analyzes the unique
changes in the current Git branch compared to the base branch
(main or master). It captures added, modified, renamed, and deleted
files for each commit in the branch. This tool helps in tracking
the evolution of the branch and facilitates better management of
code changes.

The script includes debugging utilities to assist developers in
monitoring the process. It also handles scenarios like detached
HEAD states and non-existent base branches, improving user
experience and guiding them on necessary actions.

* Refactor file change detection to use external script

This commit simplifies the workflow for detecting changed files by
replacing the inline script with a call to an external shell script
located in the `bin/actions` directory. The new approach enhances
readability and maintainability of the workflow file. The logic for
processing added, modified, and deleted files has been encapsulated
in `changed_files.sh`, which allows for easier updates in the
future without cluttering the main workflow file.

* Ignore non-content filepaths in changed_files.sh

This change adds a check to the `append_unique_existing_filepath` function
to ignore any filepaths that do not start with 'content/'. This ensures
that only relevant content files are processed, which helps to prevent
unnecessary errors and improves the performance of the script. This
modification is crucial for maintaining the integrity of the file
operations performed by this action.

* updated script to write changed and deleted filepaths to different text files; updated and old release not to test script

* restored content to match the main branch

* added error and warn functions to script

* changed function name from warning to warn

* removed improper sed based regex property

* getting rid of bash syntax in favor of pure SH

* changed file names to match current conventions; replaced old gather action with the new script

* started the integration of the deleted files into the GH workflow

* workflow was reporting that deleted-files.txt did not existl maybe this will fix that

* API-44751 Update SendToLPB Workflow

The changes made to the send-to-lpb.yaml workflow aim to
improve the handling of changed and deleted files during
the GitHub Actions execution. The primary modifications
include removing unused debug statements and updating the
deleted file handling process in the send to production
section.

The removal of debug statements enhances code readability
and maintainability by eliminating unnecessary output.
The updated deleted file handling process includes
downloading the list of deleted files as an artifact and
then sending this information to the development
environment.

* API-44751 Update SendToLPB Workflow

The changes in this commit aim to improve the workflow for sending content
to the Lighthouse Platform Backend (LPB).  The primary goal is to handle
deleted files before changed files. This alteration addresses a potential
issue where non-technical users might delete and then recreate release notes
with the same name or date through the Github UI, resulting in data loss.

In order to implement this change the steps for getting and sending deleted
files have been added prior to the steps for getting and sending changed files.
Additionally the changed_files.txt file is now being processed after the
deleted_files.txt.  This will prevent accidental deletion of newly uploaded
release notes with the same name as a previously deleted release note.

* API-44751 missed this file

* API-44751 Update SendToLPB Workflow

I'm experimenting with how to get a global raw diff between the
current feature branch in the PR and its base branch which is
assumed to be main.  When this works it will greatly simplify
the business logic in the changed_files.sh script.

* API-44751 Testing with get raw git diff inside GH actions

* API-44751 updated to get rid of old way of doing stuff

* snapshot

* zeroing in on a solution

* API-44751 Update Checkout and Diff Logic

The checkout logic has been modified to always use the main branch
and the diff calculation now writes the result directly to a file.
This allows for easier usage in subsequent steps of the workflow
and improves performance by avoiding the need to output the diff
as a step variable.

Changes have also been made to the changed_files.sh script, which
now assumes that the base branch is main when running locally, and
simplifies the logic for classifying the changed and deleted files.

* snapshot testing

* snapshot testing

* snapshot testing

* snapshot testing darn -- name things; also dump_env.sh did not have its execuatable premission set which apparently is a best practice in GH actions

* snapshot testing

* snapshot - wrong indention in yaml file

* snapshot - testing - forgot the pipe on a multi-line run block

* added content/ filter

* found another syntax error

* snapshot - trying a different way more safely filter content/

* snapshot - found an empty file problem

* snapshot - testing reveal more bash contamination in the SH script

* snapshot resorting to real basic stuff to parse a git diff raw line

* snapshot - fixed non posix compliant sh thing

* snapshot - testing rename and add inside of content

* API-44751 Remove Unused Workflow Established Consistent Pattern

The get_raw_git_diff.yaml workflow has been removed as it is no longer needed and was only used
during testing.

all changes were made to optimize performance in workflows and correcting bugs where files with spaces were not being correctly processed.  Currently we have no file paths were spaces or special
characters are used; however, if a future changed were to add these kinds of filepaths
then the existing code would break.

Updated the way that DATE_STRING was being set to ensure consistent code in all steps.
The new way of setting the DATE_STRING value is more readable / maintainable.

* snapshot - should be last test <fingers crossed>

* API-44751 changed step names to avoid confusion

* API-44751 ran npm audit fix