Update 2025-09-16.md
Updated to add notes from ux. 
Add "to learn more about this API, review the API documentation." The link leads to the prod version of the documentation that doesn't exist yet. 
Release note will not be published until the OAS is available in production. I will confirm the link is the correct one before I publish.