Create 2025-08-25.md (#277)
* Create 2025-09-16.md

Release note for launch of health care cost and coverage

* Update 2025-09-16.md

Updated to add notes from ux. 
Add "to learn more about this API, review the API documentation." The link leads to the prod version of the documentation that doesn't exist yet. 
Release note will not be published until the OAS is available in production. I will confirm the link is the correct one before I publish.

* Update and rename 2025-09-16.md to 2025-08-28.md

Release is planned for tomorrow! Just updating the release date and double checking the link goes where it should (API is currently in shadow mode).